Исправление ошибки 0 лет
