+++ Raskatova1/result.cs
@@ -1,10 +1,12 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Raskatova1
+﻿namespace Raskatova1
 {
-    
+    class Result
+    {
+        public string FullName { get; set; }
+        public int Age { get; set; }
+        public string Rank { get; set; }
+        public int TimeOfStart { get; set; }
+        public int TimeOfFinish { get; set; }
+        public int TotalTime { get; set; }
+    }
 }
